Create SlotMachine
A project that allows you to play a slot machine as you would in a casino. The program asks you how much money you are starting with and the lines you would like to bet on. There is also a "quit" function that allows you to exit the slot machine and tells you how much money you have earned at the end of your playing.